backend/pyproject.toml
[project]
name = "chief-of-staff-backend"
version = "0.1.0"
description = "FastAPI backend for Chief of Staff"
requires-python = ">=3.11"
dependencies = [
    "APScheduler>=3.10.0",
    "PyJWT[crypto]>=2.9.0",
    "asyncpg>=0.29.0",
    "edge-tts>=7.0.0",
    "email-validator>=2.0.0",
    "fastapi>=0.115.0",
    "httpx>=0.28.1",
    "langchain-core>=0.3.0",
    "langchain-mcp-adapters>=0.2.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "langsmith>=0.1.0",
    "mcp>=1.0.0",
    "openai-agents>=0.14.4",
    "pgvector>=0.3.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.2.2",
    "redis[hiredis]>=5.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    # "tensorflow==2.16.2",  # TODO: re-enable once TF supports Python 3.14
    "uvicorn[standard]>=0.32.0",
    "voyageai>=0.3.0",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main", "config"]

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "auth*", "db*", "jobs*", "lib*", "models*", "routers*", "schemas*", "services*"]
